# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelHealth API",
    "description": "Sistema de apoio a analises clinicas multimodais. Nao realiza diagnostico autonomo; toda classificacao e sujeita a revisao profissional.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analyses": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Create Analysis",
        "operationId": "create_analysis_analyses_post",
        "parameters": [
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "media"
        ],
        "summary": "List Analyses",
        "operationId": "list_analyses_analyses_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "name": "created_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "external_subject exato do profissional (filtro 'Medico').",
              "title": "Created By"
            },
            "description": "external_subject exato do profissional (filtro 'Medico')."
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data inicial (inclusiva) de criacao da analise.",
              "title": "Created From"
            },
            "description": "Data inicial (inclusiva) de criacao da analise."
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data final (inclusiva) de criacao da analise.",
              "title": "Created To"
            },
            "description": "Data final (inclusiva) de criacao da analise."
          },
          {
            "name": "patient_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro por nome do paciente (substring, case-insensitive).",
              "title": "Patient Name"
            },
            "description": "Filtro por nome do paciente (substring, case-insensitive)."
          },
          {
            "name": "patient_medical_record_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro por numero de prontuario do paciente (substring, case-insensitive).",
              "title": "Patient Medical Record Number"
            },
            "description": "Filtro por numero de prontuario do paciente (substring, case-insensitive)."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_AnalysisRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/stats": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Analysis Stats",
        "description": "Estatisticas agregadas de todas as analises da instituicao com\nconsolidacao de risco ja gravada (percentual de analises conclusivas -\nver `app.risk_consolidation.service.get_analysis_consolidation_stats`),\nusadas nos \"big numbers\" da tela de revisao da analise.",
        "operationId": "get_analysis_stats_analyses_stats_get",
        "parameters": [
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/professionals": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "List Analysis Professionals",
        "description": "Opcoes do filtro \"Medico\" do historico de analises\n(ESPECIFICACAO_FRONTEND.md secao 7.9) e do campo pesquisavel de\nfuncionario no popup de observacao clinica (matricula + nome).",
        "operationId": "list_analysis_professionals_analyses_professionals_get",
        "parameters": [
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfessionalRead"
                  },
                  "title": "Response List Analysis Professionals Analyses Professionals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/clinical-support-summary": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Generate Analysis Clinical Support Summary",
        "description": "Apoio a analise clinica assistido por LLM para esta analise\nmultimodal especifica (botao \"Analisar dados clinicos\" da tela de\nrevisao/relatorio da analise). Consolida os achados JA PRODUZIDOS pelos\nprocessadores de modalidade (imagem/audio/video/texto) e o risco JA\nCALCULADO deterministicamente em um resumo com visao clinica, causas\nprovaveis e direcionamento sugerido - sempre como apoio, nunca como\ndiagnostico ou substituicao da analise do profissional responsavel\n(ver `app.clinical_support.service.\ngenerate_analysis_clinical_support_summary`).",
        "operationId": "generate_analysis_clinical_support_summary_analyses__analysis_id__clinical_support_summary_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisClinicalSupportSummaryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analyses/{analysis_id}/media/{media_id}/confirm": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Confirm Upload",
        "operationId": "confirm_upload_analyses__analysis_id__media__media_id__confirm_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Media Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/media/local-storage/{token}": {
      "put": {
        "tags": [
          "media"
        ],
        "summary": "Upload To Local Storage",
        "operationId": "upload_to_local_storage_media_local_storage__token__put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisCreate": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "additional_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Text"
          },
          "structured_clinical_inputs": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Structured Clinical Inputs"
          }
        },
        "type": "object",
        "required": [
          "patient_id"
        ],
        "title": "AnalysisCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaConfirmRequest": {
        "properties": {
          "checksum_sha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Checksum Sha256"
          }
        },
        "type": "object",
        "required": [
          "checksum_sha256"
        ],
        "title": "MediaConfirmRequest"
      },
...
      "PageResponse_AnalysisRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRead"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "title": "PageResponse[AnalysisRead]"
      },
...
    }
  }
}